refactor: move stored record framing into module
